Eliminate Overlay.KeyBinding namespace and move everything to Settings.Section.Input
