+++ osu.Game/Overlays/Settings/Sections/Input/KeyBindingPanel.cs
@@ -4,11 +4,11 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Input.Bindings;
-using osu.Game.Overlays.KeyBinding;
 using osu.Game.Overlays.Settings;
+using osu.Game.Overlays.Settings.Sections.Input;
 using osu.Game.Rulesets;
 
-namespace osu.Game.Overlays
+namespace osu.Game.Overlays.Settings.Sections.Input
 {
     public class KeyBindingPanel : SettingsSubPanel
     {
