+++ osu.Game/Overlays/Settings/Sections/Input/RulesetBindingsSection.cs
@@ -7,7 +7,7 @@ using osu.Game.Graphics;
 using osu.Game.Overlays.Settings;
 using osu.Game.Rulesets;
 
-namespace osu.Game.Overlays.KeyBinding
+namespace osu.Game.Overlays.Settings.Sections.Input
 {
     public class RulesetBindingsSection : SettingsSection
     {
